Added assests folder to src for storing css and js data
