--- client/src/assets/graphData.js
@@ -1,32 +1,0 @@
-export var graphData = [
-    {
-        id: 1,
-        data: [
-            53838.7, 53494.99, 53638.72,
-            53431.06, 53638.04, 53064.65,
-            53010.37, 53136.36, 52038.14,
-            50357.07, 47846.62, 47711.33,
-            47347.82, 47597.01, 47574.5,
-            47481.47, 46588.66, 47172.78,
-            47476.7, 48014.55, 48279.87,
-            47910.93, 48321.62, 48467.97
-        ],
-        // last_day
-    },
-    {
-        id: 2,
-        data: [
-            58741.88, 58671.21, 57232.6,
-            56280.81, 56138.99, 57566.85,
-            56633.53, 57162.66, 58052.86,
-            58987.27, 53751.39, 53757.67,
-            54638.94, 54759.05, 54534.9,
-            57317.17, 56944.78, 57838.06,
-            57666.89, 56987.97, 57194.41,
-            57226.5, 56351.04, 56521.46,
-            57033.74, 53638.04, 46588.66,
-            48467.97
-        ]
-        // last_week
-    }
-];
