+++ client/src/components/GraphContainer.js
@@ -1,13 +1,13 @@
 import React, { useEffect } from 'react';
 // import './GraphContainer.css';
 import IndividualGraph from './Graph.js';
-import { graphData } from '../assets/graphData.js';
+import { graphData } from '../assets/js/graphData.js';
 
 const GraphContainer = () => {
     useEffect(() => {
         console.log(graphData);
+        console.log("^ GC");
     });
-
 
     return (
         <div className="container">
